Solve data base error when try to delete patient in charge of an other

+++ src/main/java/com/salesianos/triana/VaxConnectApi/user/service/PatientService.java
@@ -179,7 +179,8 @@ public class PatientService {
 
     public void deleteByPatientId (String id){
         UUID validId = UUID.fromString(id);
-        if (patientRepository.countDependentsByPatient(validId) == 0)
+        if (patientRepository.countDependentsByPatient(validId) == 0
+                && patientRepository.countPatientsInChargeOf(validId) == 0)
             patientRepository.deleteById(validId);
          else
             throw new PatientHasDependentsException();
